Switch to server-rendered home page
